Migrate form conditional case matching helper from backbone utilities
